Rearranged tests in different files; added a new test for the flooding response
