+++ src/tests.rs
@@ -1,4 +1,6 @@
+mod wg_tests;
 mod fragments;
+mod flood;
 mod simcontroller;
 
 fn run_all_tests() {
